+++ lissandra/interfaz.h
@@ -19,7 +19,7 @@
 
 //Prototipos
 char *fs_select(char *tabla, uint16_t key);
-void fs_insert(char *tabla, uint16_t key, char *value, int timestamp);
+void fs_insert(char *tabla, uint16_t key, char *value, uint64_t timestamp);
 void fs_create(char *tabla, char *tipo_consistencia, int particiones, int tiempo_compactacion);
 t_list *fs_describe(char *tabla);
 void fs_drop(char *tabla);
